requirements.txt
apache-airflow==2.10.0
apache-airflow-providers-http==4.13.0
apache-airflow-providers-sqlite==3.9.0
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.16.0
pyspark==3.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
requests==2.32.3
requests-toolbelt==1.0.0
aiohttp==3.10.5
httpx==0.27.2
python-dotenv==1.0.1
requests
apache-airflow==2.10.0
apache-airflow-providers-apache-spark==4.10.0